Split Github GraphQL to 2 functions for clarity, split Github module
